+++ modules/apps/collaboration/message-boards/message-boards-uad/src/main/java/com/liferay/message/boards/uad/display/MBMessageUADDisplay.java
@@ -14,59 +14,16 @@
 
 package com.liferay.message.boards.uad.display;
 
-import com.liferay.message.boards.model.MBMessage;
 import com.liferay.message.boards.uad.constants.MBUADConstants;
-
-import com.liferay.portal.kernel.portlet.LiferayPortletRequest;
-import com.liferay.portal.kernel.portlet.LiferayPortletResponse;
 
 import com.liferay.user.associated.data.display.UADDisplay;
 
 import org.osgi.service.component.annotations.Component;
-import org.osgi.service.component.annotations.Reference;
-
-import java.util.Locale;
-import java.util.Map;
 
 /**
  * @author Brian Wing Shun Chan
- * @generated
  */
 @Component(immediate = true, property =  {
 	"model.class.name=" + MBUADConstants.CLASS_NAME_MB_MESSAGE}, service = UADDisplay.class)
-public class MBMessageUADDisplay implements UADDisplay<MBMessage> {
-	public String getApplicationName() {
-		return MBUADConstants.APPLICATION_NAME;
-	}
-
-	public String[] getDisplayFieldNames() {
-		return _mbMessageUADDisplayHelper.getDisplayFieldNames();
-	}
-
-	@Override
-	public String getEditURL(MBMessage mbMessage,
-		LiferayPortletRequest liferayPortletRequest,
-		LiferayPortletResponse liferayPortletResponse)
-		throws Exception {
-		return _mbMessageUADDisplayHelper.getMBMessageEditURL(mbMessage,
-			liferayPortletRequest, liferayPortletResponse);
-	}
-
-	public String getKey() {
-		return MBUADConstants.CLASS_NAME_MB_MESSAGE;
-	}
-
-	@Override
-	public Map<String, Object> getNonanonymizableFieldValues(
-		MBMessage mbMessage) {
-		return _mbMessageUADDisplayHelper.getUADEntityNonanonymizableFieldValues(mbMessage);
-	}
-
-	@Override
-	public String getTypeName(Locale locale) {
-		return "MBMessage";
-	}
-
-	@Reference
-	private MBMessageUADDisplayHelper _mbMessageUADDisplayHelper;
+public class MBMessageUADDisplay extends BaseMBMessageUADDisplay {
 }
